+++ packages/quick_actions/example/android/app/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
   <application android:name="io.flutter.app.FlutterApplication"
                android:label="quick_actions_example"
                android:icon="@mipmap/ic_launcher">
-    <activity android:name=".MainActivity"
+    <activity android:name=".EmbeddingV1Activity"
               android:launchMode="singleTop"
               android:theme="@style/LaunchTheme"
               android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection"
@@ -16,12 +16,13 @@
           android:name="io.flutter.app.android.SplashScreenUntilFirstFrame"
           android:value="true"/>
     </activity>
-    <activity android:name="dev.flutter.plugins.quickactionexample.MainActivity"
+    <activity android:name=".MainActivity"
       android:launchMode="singleTop"
       android:theme="@style/LaunchTheme"
       android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection"
       android:hardwareAccelerated="true"
-      android:windowSoftInputMode="adjustResize">
+      android:windowSoftInputMode="adjustResize"
+        android:exported="true">
       <intent-filter>
         <action android:name="android.intent.action.MAIN"/>
         <category android:name="android.intent.category.LAUNCHER"/>
