--- packages/quick_actions/example/android/app/src/main/java/dev/flutter/plugins/quickactionexample/MainActivity.java
@@ -1,14 +1,0 @@
-package dev.flutter.plugins.quickactionexample;
-
-import io.flutter.embedding.android.FlutterActivity;
-import io.flutter.embedding.engine.FlutterEngine;
-import io.flutter.plugins.quickactions.QuickActionsPlugin;
-
-public class MainActivity extends FlutterActivity {
-
-  @Override
-  public void configureFlutterEngine(FlutterEngine flutterEngine) {
-    super.configureFlutterEngine(flutterEngine);
-    flutterEngine.getPlugins().add(new QuickActionsPlugin());
-  }
-}
